+++ src/main/java/roomescape/dto/ReservationDto.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dto;
 
 public record ReservationDto(String name, String date, String time) {
 }
